Replaced stupid typedef mistake u_int8 -> uint8_t and started work on CASH class

+++ DND_Character_Maker_Backend/headers/ability_score_bonus.h
@@ -5,13 +5,13 @@
 class ABILITY_SCORE_BONUS
 {
 public:
-	ABILITY_SCORE_BONUS( ABILITY_SCORE_TYPES type, u_int bonus_value );
+	ABILITY_SCORE_BONUS( ABILITY_SCORE_TYPES type, uint16_t bonus_value );
 	~ABILITY_SCORE_BONUS();
 
-	u_int get_bonus_value()				{ return m_bonus_value; };
+	uint16_t get_bonus_value()				{ return m_bonus_value; };
 	ABILITY_SCORE_TYPES get_bonus_type()	{ return m_type; };
 
 private:
 	ABILITY_SCORE_TYPES m_type;
-	u_int				m_bonus_value;
+	uint16_t				m_bonus_value;
 };
