+++ DND_Character_Maker_Backend/headers/racial_traits.h
@@ -9,15 +9,15 @@
 class RACIAL_TRAITS
 {
 public:
-	RACIAL_TRAITS( DND_RACE race, u_int age, u_int speed, string align_desc, DND_SIZE size, std::vector<DND_LANGUAGE> langs, std::vector<DND_TOOL> tool_profs,
+	RACIAL_TRAITS( DND_RACE race, uint16_t age, uint16_t speed, string align_desc, DND_SIZE size, std::vector<DND_LANGUAGE> langs, std::vector<DND_TOOL> tool_profs,
 				   std::vector<TRAIT*> traits, std::vector<ABILITY_SCORE_BONUS*> ability_bonuses, bool need_to_pick_subrace, DND_RACE main_race = DND_RACE::INVALID );
 
 	~RACIAL_TRAITS();
 
 	DND_RACE							get_race() { return m_race; };
 	DND_RACE							get_main_race() { return m_main_race; };
-	u_int								get_typical_age() { return m_typical_age; };
-	u_int								get_speed() { return m_race_speed; };
+	uint16_t								get_typical_age() { return m_typical_age; };
+	uint16_t								get_speed() { return m_race_speed; };
 	string								get_alignment_desc() { return m_alignment_description; };
 	DND_SIZE							get_size() { return m_race_size; };
 	std::vector<DND_LANGUAGE>			get_languages() { return m_race_languages; };
@@ -37,8 +37,8 @@ private:
 	DND_SIZE							m_race_size{ DND_SIZE::INVALID };
 	bool								m_must_pick_subrace{ false };
 
-	u_int								m_typical_age{ 0 };
-	u_int								m_race_speed{ 0 };
+	uint16_t								m_typical_age{ 0 };
+	uint16_t								m_race_speed{ 0 };
 
 	string								m_alignment_description{ "not set" };
